+++ test/bigtest/interactors/line-details-page.js
@@ -18,7 +18,7 @@ import { TIMEOUT } from './const';
 }
 
 @interactor class RelatedInvoicesAccordion {
-  static defaultScope = '#relatedInvoices';
+  static defaultScope = '#invoiceLines';
 
   invoices = collection('[class*=mclRow---]', {
     link: clickable('[data-test-link-to-invoice]'),
@@ -51,7 +51,7 @@ export default interactor(class LineDetailsPage {
   isLoaded = isPresent('#ItemDetails');
   relatedInvoicesAccordion = new RelatedInvoicesAccordion();
   notesAccordion = new NotesAccordion();
-  lineInvoicesIsPresent = isPresent('#lineInvoices');
+  lineInvoicesIsPresent = isPresent('#invoiceLines');
   closingReasonMessage = isPresent('[data-test-message-banner]');
   relatedAgreementLinesAccordion = new RelatedAgreementLinesAccordion();
   agreementLinesArePresent = isPresent('#po-line-agreement-lines');