+++ test/bigtest/network/factories/invoice-line.js
@@ -3,4 +3,5 @@ import faker from 'faker';
 
 export default Factory.extend({
   id: faker.datatype.uuid,
+  invoiceLineStatus: 'open',
 });